# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/calendar_event_attendees:
    get:
      tags:
        - CalendarEventAttendees
      summary: List CalendarEventAttendees
      description: Get a paginated list of CalendarEventAttendees with optional filtering and sorting
      operationId: listCalendarEventAttendees
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventAttendeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarEventAttendees
      summary: Create CalendarEventAttendee
      description: Create a new CalendarEventAttendee
      operationId: createCalendarEventAttendee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEventAttendeeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventAttendee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/calendar_event_exceptions:
    get:
      tags:
        - CalendarEventExceptions
      summary: List CalendarEventExceptions
      description: Get a paginated list of CalendarEventExceptions with optional filtering and sorting
      operationId: listCalendarEventExceptions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventExceptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CalendarEventExceptions
      summary: Create CalendarEventException
      description: Create a new CalendarEventException
      operationId: createCalendarEventException
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEventExceptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventException'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    CalendarEventAttendee:
      type: object
      required:
        - id
        - company_id
        - event_id
        - user_id
        - state
        - access_token
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        state:
          type: string
        access_token:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateCalendarEventAttendeeRequest:
      type: object
      required:
        - company_id
        - event_id
        - user_id
        - state
        - access_token
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        state:
          type: string
        access_token:
          type: string
          format: uuid
        metadata:
          type: object
...
    CalendarEventAttendeeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEventAttendee'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CalendarEventException:
      type: object
      required:
        - id
        - company_id
        - series_id
        - event_id
        - slot_start_at
        - slot_stop_at
        - kind
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        series_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        slot_start_at:
          type: string
          format: date-time
        slot_stop_at:
          type: string
          format: date-time
        kind:
          type: string
        metadata:
          type: object
...
    CreateCalendarEventExceptionRequest:
      type: object
      required:
        - company_id
        - series_id
        - event_id
        - slot_start_at
        - slot_stop_at
        - kind
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        series_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        slot_start_at:
          type: string
          format: date-time
        slot_stop_at:
          type: string
          format: date-time
        kind:
          type: string
        metadata:
          type: object
...
    CalendarEventExceptionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEventException'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []