# src/presentation/http/calendar_event_attendee_handler.rs
//! CalendarEventAttendee REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{CalendarEventAttendeeService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateCalendarEventAttendeeDto, UpdateCalendarEventAttendeeDto, PatchCalendarEventAttendeeDto, CalendarEventAttendeeResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum CalendarEventAttendeeError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for CalendarEventAttendeeError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for CalendarEventAttendeeError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "CALENDAREVENTATTENDEE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "CALENDAREVENTATTENDEE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CALENDAREVENTATTENDEE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "CALENDAREVENTATTENDEE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for CalendarEventAttendee.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /calendar_event_attendees | List with pagination |
/// | POST | /calendar_event_attendees | Create new |
/// | GET | /calendar_event_attendees/:id | Get by ID |
/// | PUT | /calendar_event_attendees/:id | Full update |
/// | PATCH | /calendar_event_attendees/:id | Partial update |
/// | DELETE | /calendar_event_attendees/:id | Soft delete |
/// | POST | /calendar_event_attendees/bulk | Bulk create |
/// | POST | /calendar_event_attendees/upsert | Upsert |
/// | GET | /calendar_event_attendees/trash | List deleted |
/// | POST | /calendar_event_attendees/:id/restore | Restore |
/// | DELETE | /calendar_event_attendees/empty | Empty trash |
/// | GET | /calendar_event_attendees/:id/deleted | Get deleted by ID |
/// | DELETE | /calendar_event_attendees/trash/:id | Permanent delete from trash |
/// | GET | /calendar_event_attendees/count | Count active entities |
/// | GET | /calendar_event_attendees/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(CalendarEventAttendeeService::with_repository(repository));
/// let router = create_calendar_event_attendee_routes(service);
/// ```
pub fn create_calendar_event_attendee_routes(service: Arc<CalendarEventAttendeeService>) -> Router {
    BackboneCrudHandler::<CalendarEventAttendeeService, CalendarEventAttendee, CreateCalendarEventAttendeeDto, UpdateCalendarEventAttendeeDto, CalendarEventAttendeeResponseDto>::routes(
        service,
        "/calendar_event_attendees",
    )
}

/// Create Axum router with only the read (GET) endpoints for CalendarEventAttendee.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_calendar_event_attendee_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_calendar_event_attendee_read_routes(service: Arc<CalendarEventAttendeeService>) -> Router {
    BackboneCrudHandler::<CalendarEventAttendeeService, CalendarEventAttendee, CreateCalendarEventAttendeeDto, UpdateCalendarEventAttendeeDto, CalendarEventAttendeeResponseDto>::read_routes(
        service,
        "/calendar_event_attendees",
    )
}

/// Create Axum router with only the write (mutation) endpoints for CalendarEventAttendee.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_calendar_event_attendee_write_routes(service: Arc<CalendarEventAttendeeService>) -> Router {
    BackboneCrudHandler::<CalendarEventAttendeeService, CalendarEventAttendee, CreateCalendarEventAttendeeDto, UpdateCalendarEventAttendeeDto, CalendarEventAttendeeResponseDto>::write_routes(
        service,
        "/calendar_event_attendees",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_calendar_event_attendee_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<CalendarEventAttendeeService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_calendar_event_attendee_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}